ISSUE #1236 - Now the model updated notifications are created correctly
